ulong instead of uint; patch by chrissalch BUG:207327
svn path=/trunk/KDE/kdeplasma-addons/; revision=1029407

+++ applets/microblog/microblog.cpp
@@ -379,9 +379,9 @@ void MicroBlog::dataUpdated(const QString& source, const Plasma::DataEngine::Dat
 
         //add the newbies
         int newCount = 0;
-        uint maxId = m_lastTweet;
+        ulong maxId = m_lastTweet;
         foreach (const QString &id, data.keys()) {
-            uint i = id.toUInt();
+            ulong i = id.toULong();
             //kDebug() << i << m_lastTweet;
             if (i > m_lastTweet) {
                 newCount++;
@@ -462,7 +462,7 @@ void MicroBlog::showTweets()
     // Add more tweetWidgets if there are not enough
 
     if (m_tweetMap.count() > m_historySize) {
-        QMap<uint, Plasma::DataEngine::Data>::iterator it = m_tweetMap.begin();
+        QMap<ulong, Plasma::DataEngine::Data>::iterator it = m_tweetMap.begin();
         while (it != m_tweetMap.end() && m_tweetMap.count() > m_historySize) {
             it = m_tweetMap.erase(it);
         }
@@ -519,7 +519,7 @@ void MicroBlog::showTweets()
     }
 
     int i = 0;
-    QMap<uint, Plasma::DataEngine::Data>::iterator it = m_tweetMap.end();
+    QMap<ulong, Plasma::DataEngine::Data>::iterator it = m_tweetMap.end();
     while (it != m_tweetMap.begin()) {
         Plasma::DataEngine::Data &tweetData = *(--it);
         QString user = tweetData.value("User").toString();
